chore: rename ration_balancer to model, more descriptive as none of this code will balance a ration
