+++ src/nasem_dairy/model/execute_model.py
@@ -23,10 +23,10 @@ import nasem_dairy.nasem_equations.unused as unused
 import nasem_dairy.nasem_equations.urine as urine
 import nasem_dairy.nasem_equations.water as water
 
-import nasem_dairy.ration_balancer.constants as constants
-import nasem_dairy.ration_balancer.input_validation as validate
+import nasem_dairy.model.constants as constants
+import nasem_dairy.model.input_validation as validate
 import nasem_dairy.model_output.ModelOutput as output
-import nasem_dairy.ration_balancer.ration_balancer_functions as ration_funcs
+import nasem_dairy.model.utilities as ration_funcs
 
 def execute_model(user_diet: pd.DataFrame,
                   animal_input: dict,
