+++ src/nasem_dairy/nasem_equations/amino_acid.py
@@ -6,7 +6,7 @@ import math
 import numpy as np
 import pandas as pd
 
-import nasem_dairy.ration_balancer.ration_balancer_functions as ration_funcs
+import nasem_dairy.model.utilities as ration_funcs
 
 def calculate_MiTPAAProf(AA_list: list, coeff_dict: dict) -> np.ndarray:
     req_coeffs = [
