+++ src/nasem_dairy/nasem_equations/body_composition.py
@@ -5,7 +5,7 @@ import math
 
 import numpy as np
 
-import nasem_dairy.ration_balancer.ration_balancer_functions as ration_funcs
+import nasem_dairy.model.utilities as ration_funcs
 
 
 def calculate_Frm_Gain_empty(Frm_Gain: float,
