+++ src/nasem_dairy/nasem_equations/energy_requirement.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-import nasem_dairy.ration_balancer.ration_balancer_functions as ration_funcs
+import nasem_dairy.model.utilities as ration_funcs
 
 def calculate_An_NEmUse_NS(An_StatePhys: str, 
                            An_BW: float,
