+++ src/nasem_dairy/nasem_equations/fecal.py
@@ -4,7 +4,7 @@
 import numpy as np
 import pandas as pd
 
-import nasem_dairy.ration_balancer.ration_balancer_functions as ration_funcs
+import nasem_dairy.model.utilities as ration_funcs
 
 
 def calculate_Fe_rOMend(Dt_DMIn: float, coeff_dict: dict) -> float:
