+++ src/nasem_dairy/nasem_equations/gestation.py
@@ -5,7 +5,7 @@ import math
 import numpy as np
 import pandas as pd
 
-import nasem_dairy.ration_balancer.ration_balancer_functions as ration_funcs
+import nasem_dairy.model.utilities as ration_funcs
 
 
 def calculate_Uter_Wtpart(Fet_BWbrth: float, coeff_dict: dict) -> float:
