+++ src/nasem_dairy/nasem_equations/infusion.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-import nasem_dairy.ration_balancer.ration_balancer_functions as ration_funcs
+import nasem_dairy.model.utilities as ration_funcs
 
 def calculate_Inf_TPIn(Inf_CPIn: float, Inf_NPNCPIn: float) -> float:
     Inf_TPIn = Inf_CPIn - Inf_NPNCPIn  # Line 848
