+++ src/nasem_dairy/nasem_equations/urine.py
@@ -3,7 +3,7 @@
 import numpy as np
 import pandas as pd
 
-import nasem_dairy.ration_balancer.ration_balancer_functions as ration_funcs
+import nasem_dairy.model.utilities as ration_funcs
 
 
 def calculate_Ur_Nout_g(Dt_CPIn: float,
